@@ -22,44 +22,44 @@
     "start": "ng serve",
     "build:dev": "ng build",
     "build:dev:watch": "npm run build:dev && ng build --watch",
-    "build:prod": "ng build --prod --aot --extract-css",
+    "build:prod": "ng build --configuration production --aot",
     "lint:sass": "sass-lint -c .sass-lint.yml Client/**/*.scss --verbose",
     "lint:ts": "tslint -c tslint.json Client/**/*.ts"
   },
   "dependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/platform-server": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular/animations": "13.3.1",
+    "@angular/common": "13.3.1",
+    "@angular/compiler": "13.3.1",
+    "@angular/core": "13.3.1",
+    "@angular/forms": "13.3.1",
+    "@angular/platform-browser": "13.3.1",
+    "@angular/platform-browser-dynamic": "13.3.1",
+    "@angular/platform-server": "13.3.1",
+    "@angular/router": "13.3.1",
     "@microsoft/signalr": "3.0.1",
-    "@ng-bootstrap/ng-bootstrap": "5.2.1",
+    "@ng-bootstrap/ng-bootstrap": "12.0.1",
+    "@popperjs/core": "2.11.4",
+    "acorn": "^6.0.0",
+    "acorn-dynamic-import": "4.0.0",
     "bootstrap": "4.4.1",
     "core-js": "^3.0.0",
     "file-loader": "2.0.0",
     "font-awesome": "4.7.0",
-    "isomorphic-fetch": "2.2.1",
-    "jquery": "3.4.1",
-    "ngx-toastr": "10.1.0",
+    "isomorphic-fetch": "^3.0.0",
+    "jquery": "^3.6.0",
+    "ngx-toastr": "12.1.0",
     "normalize.css": "8.0.0",
-    "@popperjs/core": "2.0.0",
-    "rxjs": "~6.4.0",
+    "popper.js": "1.16.1",
+    "rxjs": "~6.6.7",
     "rxjs-compat": "~6.4.0",
-    "webpack-dev-server": "3.1.14",
-    "zone.js": "~0.9.1",
-    "acorn-dynamic-import": "4.0.0",
-    "acorn": "^6.0.0",
-    "popper.js": "1.16.1"
+    "webpack-dev-server": "4.7.4",
+    "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.803.23",
-    "@angular/cli": "8.3.23",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "13.3.1",
+    "@angular/cli": "13.3.1",
+    "@angular/compiler-cli": "13.3.1",
+    "@angular/language-service": "13.3.1",
     "@types/core-js": "2.5.0",
     "@types/hammerjs": "2.0.35",
     "@types/jasmine": "^3.3.12",
@@ -67,27 +67,27 @@
     "@types/protractor": "4.0.0",
     "@types/selenium-webdriver": "3.0.10",
     "codelyzer": "^5.0.0-beta.1",
-    "eslint": "4.18.2",
-    "handlebars": "^4.7.2",
+    "eslint": "^8.10.0",
+    "handlebars": "^4.7.7",
     "jasmine-core": "^3.3.0",
     "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^4.0.1",
+    "karma": "^6.3.17",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^2.0.0",
     "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
     "lodash": "^4.17.15",
-    "merge": "1.2.1",
-    "npm-watch": "0.5.0",
-    "protractor": "^5.4.2",
-    "rxjs-tslint": "^0.1.7",
-    "sass-lint": "1.12.1",
+    "merge": "2.1.1",
+    "npm-watch": "^0.11.0",
+    "protractor": "^7.0.0",
+    "rxjs-tslint": "^0.1.8",
+    "sass-lint": "^1.13.1",
     "ts-helpers": "1.1.2",
     "ts-node": "~7.0.1",
-    "tslint": "^5.14.0",
-    "typedoc": "^0.15.0",
-    "typescript": "3.4.5",
+    "tslint": "^6.1.3",
+    "typedoc": "^0.22.14",
+    "typescript": "4.6.3",
     "url-loader": "1.1.1",
-    "webpack": "^4.29.6"
+    "webpack": "^5.72.0"
   }
 }
