@@ -57,7 +57,7 @@ import { UppercasePipe } from './pipes/uppercase.pipe';
     ]
 })
 export class SharedModule {
-    static forRoot(): ModuleWithProviders {
+    static forRoot(): ModuleWithProviders<SharedModule> {
         return {
             ngModule: SharedModule,
             providers: [
